backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
pydantic-settings==2.3.4
cairosvg==2.7.1
svgwrite==1.4.3
python-multipart==0.0.9
Jinja2==3.1.4
pytest==8.3.2
httpx==0.27.2
python-jose[cryptography]==3.3.0
requests==2.32.3
